👷 (Jenkins) 更新打包目录

+++ Jenkinsfile
@@ -5,6 +5,8 @@ pipeline {
     }
     environment {
         ONLINE_SITE = 'https://www.abyssal.site/vuetify-boilerplate'
+        AUTHOR_EMAIL = 'hongxin.tang@hotmail.com'
+        DEPLOY_DIR   = '/www/wwwroot/visual'
     }
     stages {
         stage('Build') { 
@@ -15,11 +17,16 @@ pipeline {
         }
         stage('Deploy') {
             steps {
-                dir('/usr/share/nginx') {
-                    sh 'rm -rf vuetify-boilerplate.back'
-                    sh 'mv vuetify-boilerplate vuetify-boilerplate.back' 
-                    sh 'mv ${WORKSPACE}/dist .'
-                    sh 'mv dist vuetify-boilerplate'
+                dir("${DEPLOY_DIR}") {
+                    script {
+                        if (fileExists('vuetify-boilerplate.back')) {
+                            sh 'rm -rf vuetify-boilerplate.back'
+                        }
+                        if (fileExists('vuetify-boilerplate')) {
+                            sh 'mv vuetify-boilerplate vuetify-boilerplate.back'
+                        }    
+                    }
+                    sh 'mv ${WORKSPACE}/dist ./vuetify-boilerplate'
                 }
             }
         }
@@ -29,7 +36,7 @@ pipeline {
             emailext body: "View on ${ONLINE_SITE}, See detail at ${BUILD_URL}",
                     recipientProviders: [developers(), requestor()],
                     subject: "Jenkins: ${JOB_NAME} ${GIT_BRANCH} build ${currentBuild.result}",
-                    to: 'hongxin.tang@hotmail.com'
+                    to: "${AUTHOR_EMAIL}"
         }
     }
 }
